test(site): make page URL reactivity test actually exercise reactivity
The previous test mocked $app/state as a static plain object and used the
legacy `card=` query params, but the component restores cards from the
bitfield-encoded `s` param via resolveCardsFromUrl. Both the wrong param and
the wrong queried role meant the assertions matched zero cards no matter what,
and the second test's waitFor held only commented-out expectations, so it
passed unconditionally.

Rename to page.reactivity.svelte.test.ts to enable runes, back the page mock
with $state so reassigning page.url notifies the component, and assert the
restore effect both seeds cards on mount and re-runs when the URL changes.
