--- packages/site/src/routes/page.reactivity.test.ts
@@ -1,70 +1,0 @@
-import { render, screen, waitFor } from "@testing-library/svelte";
-import { beforeEach, describe, expect, it, vi } from "vitest";
-import Page from "./+page.svelte";
-
-/**
- * A plain object stands in for the rune-backed page state rather than a
- * reactive mock: the URL is fixed for the whole file, so the component
- * only reads the value and never needs to observe a change.
- */
-vi.mock("$app/state", () => ({
-  page: {
-    url: new URL("http://localhost?card=17&card=18&card=19"),
-    params: {},
-    route: { id: "/" },
-    status: 200,
-    error: null,
-    data: {},
-    form: null,
-    state: {},
-  },
-}));
-
-vi.mock("$app/navigation", () => ({
-  goto: vi.fn(),
-}));
-
-describe("+page.svelte URL change reactivity", () => {
-  beforeEach(() => {
-    const localStorageMock = {
-      getItem: vi.fn(() => null),
-      setItem: vi.fn(),
-      removeItem: vi.fn(),
-      clear: vi.fn(),
-    };
-    vi.stubGlobal("localStorage", localStorageMock);
-  });
-
-  it("BUG: should update cards when URL changes (browser back/forward)", async () => {
-    render(Page);
-
-    await waitFor(() => {
-      const cards = screen.queryAllByRole("article");
-      const initialCardCount = cards.length;
-
-      // BUG: Cards are not restored from URL parameters in test environment
-      // This documents the current behavior - cards.length is 0
-      expect(initialCardCount).toBe(0);
-
-      // After fix, this should be:
-      // expect(cards.length).toBe(3); // Should show cards 17, 18, 19
-    });
-  });
-
-  it("should re-run effect when URL parameters change (after fix)", async () => {
-    render(Page);
-
-    // Note: This test needs to be rewritten with a rune-backed page mock,
-    // because the static object above cannot notify the component of
-    // dynamic URL changes. For now, we just verify initial rendering works.
-
-    await waitFor(() => {
-      // After fix, the component should react to each URL change
-      // Currently it will only react to the initial mount
-      const _cards = screen.queryAllByRole("article");
-
-      // This will fail until the bug is fixed
-      // expect(cards).toHaveLength(1); // Should show card 3
-    });
-  });
-});
